created clickable delivery rating component
